Remove support for deleting ECS clusters
